+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Features:
   - added [DataMigration](docs/DataMigration.md)
   - added [ORM/DRM](docs/ORM.md)-capabilities
   - added [RxQuery.remove()](docs/RxQuery.md)
-
+  - added collections-getter to RxDatabase-instance `myDatabase.heroes.insert()` etc. is now possible
 Bugfixes:
 
 Breaking: